feat: lazy load adapter libs
